refactor: rename component to match Figma
